Add HoundCI configuration
HoundCI is a service which checks Pull Requests for style violations.
More information can be found here: https://houndci.com/

This file configures Hound to behave in such a way that it conforms to
the thoughtbot style guide [style-guide].

The contents of the file are a straight copy of the default
configuration of Hound (at the time of committing) [hound-config].
This means that while the inclusion of the file is not necessary at the
moment, having it around allows us to configure Hound's behaviour, if
such a need arises.

  [style-guide]: https://github.com/thoughtbot/guides/tree/master/style
  [hound-config]: https://github.com/thoughtbot/hound/blob/7535d49de6e60302901ab2646233444922cabea9/config/rubocop.yml