Send in RUBY_VERSION.dup to Gem::Version
It seems in versions of Ruby prior to 2.x, calling `Gem::Version.new`
with a frozen string as an argument will result in a RuntimeError, since
the library will attempt to mutate the string.

In versions of Ruby above 2.x the Gem::Version library does a .dup on
the object that comes in right away, so there it shouldn't be a problem.