CI: add CircleCI config
TravisCI performance isn't great, and I need to be able to SSH into the
build in order to debug. I tried the `curl` command [recommended in the
Travis docs][1], but that didn't do the trick. Circle doesn't support
matrix the way that Travis does, but I figure we can manage with a
slightly more verbose config file.

- Needed to add `-stretch` to be able to install ghostscript for fonts.
  By default CircleCI uses debian `jessie`, a previous version, which
  has an older version of `ghostscript`, at `9.06`, and I was unable to
  install it due to package dependency issues. CircleCI recommends
  [pinning to an OS version as a best practice][2] anyway.

[1]: https://docs.travis-ci.com/user/running-build-in-debug-mode/#restarting-a-job-in-debug-mode-via-api
[2]: https://circleci.com/docs/2.0/circleci-images/#best-practices