AcquireDrawInfo() instead of magick_malloc()-ing some memory (#314)
There might not be enough memory later if ImageMagick changes the struct's size. The size of the allocated memory is set at compile time and no other structure initialization is made when using `magick_malloc()`. This can break when the size of the structure changes with a shared library update or when ImageMagick starts relying on initializing the info structure.