Lock down development dependency versions (#355)
Because we don't check in `Gemfile.lock`, it's good to lock these down
in our `gemspec` file. This makes sure the build stays consistent if,
say, there's a major breaking release of RSpec, or once we've added
Rubocop to the build we don't want it to randomly fail every time they
add new rules.