Use a different method to get the compiler flags for ImageMagick. (#487)
Use a different method to get the compiler flags for ImageMagick

This fixes an issue when both ImageMagick 6 and 7 can be found with pkg-config. This changes the method that is used to determine the version of ImageMagick. It searches for ImageMagick-6* instead of Wand or MagickWand. This fixes an issue when both ImageMagick 6 and 7 can be found with pkg-config. And it also makes sense to remove Wand because that part of ImageMagick is not used by this library.