Fix rmfill leaks (#528)
This is a draft of changes that can be made after #523 has been merged. I created this draft PR so it can already be reviewed. Would love to get some feedback on the naming of the methods and enumeration. These methods have already been mentioned in #325. This PR includes the new methods and an example of how they can be used. I don't mind splitting this up in two pull requests but combining them will give a clearer view of how these methods will be used.