Removed automatic allocation of the ImageInfo that is broken. (#547)
This PR removes the automatic allocation of the ImageInfo that is broken. The my_info is created in case info is NULL but this is not used in the call to rm_acquire_image. And all the calls to this method are also from places where the value cannot be null. I have also removed the RB_GC_GUARD because we return the value.

And it looks like we could move this method to the rminfo.c file but that should be something for another PR.