Fix memory leak (#665)
This PR includes a fix for a memory leak inside the `Image_constitute` method. `xfree(pixels.v)` should be called before an exception is raised. This PR also includes some minor refactoring to make it more in the style of the rest of the project. And `exception = &new_image->exception;` is used to prepare for ImageMagick 7.