Deprecated the use of an unnamed argument for the opacity value of Image#transparent (#695)
To have a smooth transition from ImageMagick 6 to ImageMagick 7 some of the arguments need to change from an opacity value to an alpha value. And after a chat @mockdeep and me came up with the following idea make sure the user will know that this is happening:

Instead of only a `Quantum` value

```ruby
img.transparent('white', Magick::TransparentOpacity)
```

A named argument should be used.

```ruby
img.transparent('white', alpha: Magick::TransparentAlpha)
```

If only a value is passed in the value will be changed from opacity to alpha and a warning will be raised to inform the user that they should use a named argument and switch from opacity to alpha.