Deprecated the use of an unnamed argument for the opacity value of Image#transparent_chroma. (#722)
To have a smooth transition from ImageMagick 6 to ImageMagick 7 some of the arguments need to change from an opacity value to an alpha value. Instead of only a Quantum value a named argument should be used. If only a value is passed in the value will be changed from opacity to alpha and a warning will be raised to inform the user that they should use a named argument and switch from opacity to alpha.