Use the `magick_module` define to get the `module` field. (#797)
This is for a recent change that was made to ImageMagick 6 and 7. The library had issues to be compiled with C++20 with modules enabled (ImageMagick/ImageMagick#1650). To fix this issue a define was intruduced to resolve that problem by renaming the field in the struct but the code that uses this field should use magick_module to get the value of that field. This PR adds the define for older version of ImageMagick and uses that define to get the value of the field.