try removing Mutex
We had [a problem against Ruby 2.2][1] where setting `trace_proc` on an
image caused a segfault in certain situations. A [mutex was added][2] to
remedy this issue. I'm curious whether this is still an issue, or
whether we can remove the Mutex. One thing to note is that a reference
to this mutex is not preserved, so it isn't preventing any
multi-threading issues where multiple threads set `trace_proc`. It
appears to be used for its side effects rather than its purpose, so if
there is still a problem with this segfault, I'm wondering if there is a
more direct approach.

[1]: https://travis-ci.org/rmagick/rmagick/jobs/60001052
[2]: https://github.com/rmagick/rmagick/pull/194