switch from Test::Unit to Minitest
This changes the existing unit test framework from `Test::Unit` to
`Minitest`. The primary reason for this is to ease the transition to
RSpec, since `Minitest` [has a spec syntax][1] not far off from that of
`RSpec`. This should allow for a faster swap in terms of test syntax,
and easier diffs as we can make simple, sweeping changes.

I added a few aliases here as Minitest has fewer assertions. Since these
will be switching to the RSpec `expect` syntax before long, I don't
think it's worth the effort to switch them at the call site now, and the
aliases help reduce the diff as well.

In subsequent commits, I plan to start adding new matcher aliases that
map to the `RSpec` syntax and changing the file structures to the
`Minitest` spec syntax. I have been refactoring the tests as I've
converted them to `RSpec`, but I think for these changes I may leave the
refactoring until the end of the transition.

[1]: https://github.com/seattlerb/minitest#specs