Replace `assert_equal` with RSpec style matcher
This demonstrates the approach I want to take with converting tests to
RSpec. Rather than trying to convert tests file-by-file, we can sweep
through all of the tests and replace the matchers with the RSpec syntax.
Subsequent commits will continue to replace the existing `assert` style
matchers with `expect`. Once that is finished, we can swap in `minitest`
`spec` style, then move everything to the `spec/` folder, and finally
split the tests up and refactor them in place.