Replace `assert_block` with RSpec style matcher
This continues transitioning matcher styles to RSpec style, which should
greatly speed up the transition to RSpec.

In this case I replaced `assert_block` calls with a handful of other
matchers, mostly `to be(true)` and `to be(false)`. I also added a
`respond_to` matcher. There is an `all` matcher, but I didn't want to
get down the rabbit hole of trying to implement that, so I left the
`all?` with a `be(true)` matcher as well.

https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/respond-to-matcher
https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/all-matcher