Replace remaining assertions with RSpec style matcher
This changes `assert`, `assert_not_nil`, `assert_true`, and
`assert_false` assertions to use mostly `be(true)`, `be(false)` and
`be(nil)` checks, with the introduction of `have_key` and `include`
checks as well.

This should be the last step in switching the matchers to use RSpec
syntax. Next up is switching the test files to use `spec` syntax.

https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/predicate-matchers#should-have-key-(based-on-hash#has-key?)
https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/include-matcher