Replace `assert_nil` with RSpec style matcher (#838)
This continues transitioning matcher styles to RSpec style, which should
greatly speed up the transition to RSpec.

In this case I decided to use `to be(nil)` rather than `to be_nil`. Both
are available in RSpec, and they are essentially equivalent. The reason
I chose `be(nil)` was to be syntactically consistent with `be(true)` and
`be(false)`, since there is no longer a `be_true` in RSpec, only a
`be_truthy` which is less strict about the return value. In this latter
case I prefer to avoid `be_truthy` and be more explicit about what is
expected, either via `be(true)` or `eq(something_more_specific)`.

https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/equality-matchers#compare-using-be-(equal?)
https://relishapp.com/rspec/rspec-expectations/v/3-9/docs/built-in-matchers/be-matchers#be-nil-matcher