RSpec: move all tests to spec directory
This converts all of the tests to RSpec and removes the `minitest`
dependency. It was largely a painless thanks to the syntax having
already been converted. The only somewhat confusing change was in a
couple of places where calling `_tmpnam_` and `_dummy_img_` were
expected to raise errors on the first call but they didn't, presumably
because they had been defined by previous test calls. Since these are
private we may end up wanting to remove the tests for them at some
point, but for now I just commented out the assertion in each that
expected an error to be thrown.

After this I'll break up the tests into files by class/method and
finally split up the assertions under better descriptions.