Rubocop: enable ExpectActual rubocop rule
The `ExpectActual` rule doesn't like a constant being passed into
`expect` while a variable is passed into the test predicate. It wants
the unknown value to be passed to `expect`. It's a bit awkward in this
situation since there's no `be_included_in` matcher. However, I used
this as an excuse to refactor the test for clarity, so the violation is
gone now.