RSpec: clean up SimpleCov config
We don't really need the `.simplecov` file, since we aren't configuring
`SimpleCov` for multiple test frameworks, just RSpec. I've added several
groupings here, including the `extconf.rb` file, as well as for the
`rvg/` folder. I'm not actually sure what `RVG` does, but seems
important to have it tested as long as we have it in our code. I also
optimistically added groupings for the examples folders in the hopes
that someday we'll put those under test as well.

I also added the `spec/` directory. This is unconventional, but we have
a lot of conditions in our tests based on various configurations,
including ImageMagick version, as well as tests that are currently being
skipped. I think it would be worthwhile to track this and be aware of
which tests aren't running for one reason or another. Down the road it
would be nice to figure out how to combine coverage results across IM
versions to ensure that all branches are hit.

Finally, I added requires for `extconf.rb` and `rvg/rvg.rb`. If the
files aren't required, SimpleCov doesn't track files that aren't
required. As such, the examples folders show 100% with no files in them.
Hopefully, in the not too distant future we can add tests wrapping these
up.