Rubocop: configure Style/TrailingCommaInArguments
This changes the setting to require trailing commas in method arguments.

The primary benefit of this approach is that it reduces diff noise when
we add/remove an item from a list:

```diff
 {
-  a: 'a'
+  a: 'a',
+  b: 'b'
 }
```

vs:

```diff
 {
   a: 'a',
+  b: 'b',
 }
```

It also preserves the `git blame` on the previous line. In addition, it
has some benefits in terms of developer ergonomics. You no longer have
to remember the comma when you add/remove an item, copy/paste, or sort a
list. It's a minor tripping point, but I often have this sort of problem
working in codebases where the trailing comma is not used. Then I don't
find out about it until the tests run, or worst case it makes it into
production and we get a syntax error.

https://rubocop.readthedocs.io/en/stable/cops_style/#styletrailingcommainarguments