RSpec: enable DescribedClass rubocop rule (#962)
This rule requires us to use the `described_class` test helper rather
than explicitly naming the class being tested. The benefit of using
`described_class`, aside from being a little shorter in many cases, is
that it helps prevent false positives when we copy and paste tests if we
forget to change the class name under test.