Rubocop: configure Layout/MultilineAssignmentLayout
This rule dictates how we should align assignment across multiple lines.
The default wants us to indent the block on the next line, like so:

```rb
foo =
  if expression
    'bar'
  end
```

There is a configuration for `same_line` which would instead expect us
to format our code like so:

```rb
foo = if expression
  'bar'
end
```

I kind of prefer the first option, but all of our code is in line with
the second option, `same_line`, so I went with that instead. You can see
the impact to our codebase if we were to use the first option in
https://github.com/rmagick/rmagick/pull/971.

https://rubocop.readthedocs.io/en/stable/cops_layout/#layoutmultilineassignmentlayout