CI: set up Github Actions (#984)
This adds a new `ci.yml` to run our tests and lints on Github actions.
It also removes the CircleCI configurations. Github Actions ended up
being much cleaner, and the build runs *very* quickly as it runs all
jobs simultaneously. The only downside seems to be that as soon as one
job fails, it cancels all of the other jobs.

I used [`eregon/use-ruby-action`][eregon] rather than the default Ruby
setup action because we need support for Ruby 2.3, which isn't available
in the other Ruby actions.

It looks like we could use this to run our Windows test as well and
remove Appveyor.

[eregon]: https://github.com/eregon/use-ruby-action