CI: don't specify ruby patch version (#986)
I'd like to make the builds required, but it's kind of tricky if the
patch version changes, since each version check is a separate build
requirement. Removing the patch version for Ruby at least makes that a
little less painful. It'll still be a problem with ImageMagick patch
versions, but I'll continue to think on it.