CI: use object syntax in matrix (#987)
This changes the `matrix` options for `imagemagick-version` to be
structured as an object. The goal of using this approach is that the job
name can be consistent when we increase the patch or build numbers. This
will allow us to make the build steps required without needing to update
the required steps every time we update them.