RSpec: number repeated test variables (#980)
This adds a number in situations where we assign a new variable based on
an instance variable with the same name. So something like:

```rb
list = @list.clone
```

changes to:

```rb
list2 = @list.clone
```

The reason for this, aside from greater clarity, is that I want to
remove the instance variables. Instead, we will instantiate them locally
in the test, or use `let` blocks where appropriate.