RSpec: eliminate `let` blocks (#1025)
I've been moving in the  direction of avoiding `let` blocks as they
create indirection in the tests that can make it hard to fully
understand what the test is testing. Instead, I will err on the side of
a little more duplication in the tests. In cases where there is a lot of
setup logic to use over and over, I will often pull it out into helper
methods either in the specific test file, or in a higher level file for
use across multiple specs. More detailed argument against `let` blocks
in [this thoughtbot article][1].

[1]: https://thoughtbot.com/blog/lets-not