RSpec: remove `not_to raise_error` blocks (#1065)
This removes most multiline `not_to raise_error` blocks. These are
unnecessary as it will fail the test if an error is raised either way. I
left in some blocks in cases where the test was more focused on whether
or not it raised an error. Eventually it would be nice if those changed
to more precise tests validating the outcome rather than just checking
whether a method call raised an error.