CI: add dependency on linux build for windows
I posted in the Github Actions forums and [it was suggested][1] that the
reason our build was flaky was because we're queueing too many jobs at
once. I'm not sure why that would be a problem, but thought I'd give it
a try to see if it resolves our issues.

[1]: https://github.community/t5/GitHub-Actions/Checks-hang-with-no-output/m-p/46689#M6582