Fix bug of signal handling internally
The signal handler settings have a system-wide impact.
It seems that in some cases the app won't work properly if it just restore the signal handling to the default.

At least, the gem of CarrierWave gem is failed with RMagick 4.1.0.

To make more safety, this patch restores the handlers that had been configured instead of restoring the defaults.