Also set the alpha value of the target pixel (#1299)
The alpha channel of the target pixel should be set to make sure a proper match can be found. This fixes the issue reported in #1296. And the alpha_trait of the fill pixel should also be set to make sure the alpha channel of the resulting image will be enabled.