Added codespaces configurations (#1341)
This adds codespaces configuration so that someone can make changes to our code and run the unit tests in the browser without needing to install anything themselves. This is done automatically but the user will need to wait for this. This adds two configurations one for ImageMagick 6 and another one for ImageMagick 7. But both configurations are using Ruby 3.1.2. We cannot use the pre-build configuration because we are using a custom configuration but I hope that will become available in the future so users don't need to wait for this.